requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
pandas==1.5.3

pillow=9.3.0

torchvision==0.15.2

torch==2.0.1

matplotlib==3.7.2

tqdm=4.65.0

opencv-python==4.8.0.74